requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
pillow==10.3.0
tqdm==4.66.4
pandas==2.2.2
matplotlib==3.9.1
jupyter==1.0.0